Правка тестов

+++ pyback/tests/detections.py
@@ -1,6 +1,6 @@
 import tempfile # можно юзать также io.StringIo (более легкий модуль)
 import tarfile
-from helpers.others import downloadAndMove
+from helpers.net import downloadAndMove
 import unittest
 from os.path import join
 from settings import Settings as cfg
@@ -11,26 +11,31 @@ from neural_network.maskCNN import Mask
 
 class DetectionsTest(unittest.TestCase):
     dirName = "detections"
-    cacheDirectory = False
+    cacheDirectory = True
 
     def setUp(self):
         # скачаем тестовую обстановку
-        link = "https://vk.com/doc84996630_511676458?hash=e915563b619b68d654&dl=c0802c3ea67efa693e"
+        link = "https://vk.com/doc84996630_511877903?hash=b0be058a17a0081383&dl=4ae52327d30cae1c59"
         name = f"{self.dirName}.tar.gz"
-        downloadAndMove(link, join(cfg.TEST_IMAGE_DIR, name))  # на этом этапе мы не знамем навзание фала по ссылке
         archivePath = join(cfg.TEST_IMAGE_DIR, name)
-        tar = tarfile.open(archivePath, 'r')
-        tar.extractall(cfg.TEST_IMAGE_DIR)
-        os.remove(archivePath)
+        if not self.cacheDirectory:
+            downloadAndMove(link, join(cfg.TEST_IMAGE_DIR, name))  # на этом этапе мы не знамем навзание фала по ссылке
+            tar = tarfile.open(archivePath, 'r')
+            tar.extractall(cfg.TEST_IMAGE_DIR)
+            os.remove(archivePath)
 
     def test(self):
         mask = Mask()
         with open(os.path.join(cfg.TEST_IMAGE_DIR, 'detections', 'detections.json')) as json_file:
             data = json.load(json_file)
 
-        for filename in os.listdir(os.path.join(cfg.TEST_IMAGE_DIR, self.dirName)):
-            _f,  arg = mask.pipeline(filename, None)
-            self.assertEqual( arg, data[filename])
+        for i, filename in enumerate(os.listdir(os.path.join(cfg.TEST_IMAGE_DIR, self.dirName))):
+            filepath = os.path.join(cfg.TEST_IMAGE_DIR, self.dirName, filename)
+            _f,  arg = mask.pipeline(filepath, os.path.join(cfg.TEST_IMAGE_DIR, 'detectionsOutput', filename))
+            print(arg)
+            print(data[filename])
+            print(filename)
+            self.assertEqual(arg, data[filename])
 
     def tearDown(self):
         if not self.cacheDirectory:
